Feat/stock date management - next, finish, result

+++ src/main/java/com/DahnTa/dto/MarketPrices.java
@@ -3,9 +3,9 @@ package com.DahnTa.dto;
 import lombok.Builder;
 
 @Builder
-public record MarketPrices(int marketPrice) {
+public record MarketPrices(double marketPrice) {
 
-    public static MarketPrices create(int marketPrice) {
+    public static MarketPrices create(double marketPrice) {
 
         return MarketPrices.builder()
             .marketPrice(marketPrice)
